backend/ai-training/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.1
Pillow==10.0.0
ultralytics==8.0.196
scikit-learn==1.2.2
matplotlib==3.7.1
shap==0.42.1
